rename ItemCatalog -> CatalogItemsList
