+++ react-native/src/components/CatalogItemsList.js
